Supplier Matching/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pandas==2.2.2
scikit-learn==1.4.2
pdfplumber==0.10.3
deep-translator==1.11.4
pyquery==2.0.0
requests==2.31.0
lxml==5.1.0
openpyxl==3.1.2
openai==0.28.0
python-dotenv==1.0.1
transformers==4.40.0
torch==2.3.0
sentence-transformers==2.6.1